requirements.txt
tensorflow-gpu>=2.3.0
bert-for-tf2>=0.14.7
fasttext==0.9.2
torch>=1.7.1
# 剩余依赖自行安装